+++ dpx/src/dpx_dvi.rs
@@ -518,14 +518,16 @@ unsafe fn read_font_record(tex_id: u32) {
     let directory = new(
         ((dir_length + 1i32) as u32 as u64).wrapping_mul(::std::mem::size_of::<i8>() as u64) as u32
     ) as *mut i8;
-    if ttstub_input_read(handle.0.as_ptr(), directory, dir_length as size_t) != dir_length as i64 {
+    if ttstub_input_read(handle.0.as_ptr(), directory, dir_length as size_t) != dir_length as isize
+    {
         panic!(invalid_signature);
     }
     *directory.offset(dir_length as isize) = '\u{0}' as i32 as i8;
     free(directory as *mut libc::c_void);
     let font_name = new(((name_length + 1i32) as u32 as u64)
         .wrapping_mul(::std::mem::size_of::<i8>() as u64) as u32) as *mut i8;
-    if ttstub_input_read(handle.0.as_ptr(), font_name, name_length as size_t) != name_length as i64
+    if ttstub_input_read(handle.0.as_ptr(), font_name, name_length as size_t)
+        != name_length as isize
     {
         panic!(invalid_signature);
     }
@@ -554,7 +556,7 @@ unsafe fn read_native_font_record(tex_id: u32) {
     let font_name =
         new(((len + 1i32) as u32 as u64).wrapping_mul(::std::mem::size_of::<i8>() as u64) as u32)
             as *mut i8;
-    if ttstub_input_read(handle.0.as_ptr(), font_name, len as size_t) != len as i64 {
+    if ttstub_input_read(handle.0.as_ptr(), font_name, len as size_t) != len as isize {
         panic!(invalid_signature);
     }
     *font_name.offset(len as isize) = '\u{0}' as i32 as i8;
